Move head and tail to IndexedFrame
